Moved all the OrderHistory-related files to the folder order_history_use_case
