+++ src/main/java/order_history_use_case/OrderHistoryInteractor.java
@@ -1,6 +1,5 @@
-package order_use_case;
+package order_history_use_case;
 
-import entities.Order;
 import entities.User;
 
 import java.io.*;
