+++ src/main/java/order_history_use_case/OrderHistoryView.java
@@ -1,4 +1,4 @@
-package order_use_case;
+package order_history_use_case;
 
 public class OrderHistoryView {
     // TODO: open a new screen that displays every past order by the user in order by orderID
